Decouple some components
